+++ src/com/pmnm/roy/ui/menu/extensions/ExitPopup.java
@@ -71,7 +71,7 @@ public final class ExitPopup extends DoaObject implements IRoyContainer, Observe
 	public class Renderer extends DoaRenderer {
 
 		private Font font;
-		private int stringWidth;
+		private float stringWidth;
 		private DoaVector textDimensions = new DoaVector(576, 40);
 
 		@Override
@@ -85,7 +85,7 @@ public final class ExitPopup extends DoaObject implements IRoyContainer, Observe
 
 				FontMetrics fm = DoaGraphicsFunctions.getFontMetrics(font);
 				stringWidth = fm.stringWidth(areYouSureText);
-				int[] strSize = DoaGraphicsFunctions.unwarp(stringWidth, 0);
+				float[] strSize = DoaGraphicsFunctions.unwarp(stringWidth, 0);
 				stringWidth = strSize[0];
 			}
