+++ src/com/pmnm/roy/ui/menu/extensions/InputPopup.java
@@ -82,7 +82,7 @@ public final class InputPopup extends DoaObject implements IRoyContainer, IRoyIn
 	public class Renderer extends DoaRenderer {
 
 		private Font font;
-		private int stringWidth;
+		private float stringWidth;
 		private DoaVector textDimensions = new DoaVector(576, 40);
 
 		@Override
@@ -96,7 +96,7 @@ public final class InputPopup extends DoaObject implements IRoyContainer, IRoyIn
 
 				FontMetrics fm = DoaGraphicsFunctions.getFontMetrics(font);
 				stringWidth = fm.stringWidth(hostIDText);
-				int[] strSize = DoaGraphicsFunctions.unwarp(stringWidth, 0);
+				float[] strSize = DoaGraphicsFunctions.unwarp(stringWidth, 0);
 				stringWidth = strSize[0];
 			}
